+++ Source/BSGOutOfMemoryWatchdog.m
@@ -13,6 +13,7 @@
 #import "Bugsnag.h"
 #import "BugsnagSessionTracker.h"
 #import "Private.h"
+#import "BugsnagErrorTypes.h"
 
 @interface BSGOutOfMemoryWatchdog ()
 @property(nonatomic, getter=isWatching) BOOL watching;
@@ -199,7 +200,7 @@
             BOOL sameVersions = [lastBootOSVersion isEqualToString:osVersion] &&
                                 [lastBootBundleVersion isEqualToString:bundleVersion] &&
                                 [lastBootAppVersion isEqualToString:appVersion];
-            BOOL shouldReport = (config.enabledErrorTypes & BSGErrorTypesOOMs)
+            BOOL shouldReport = (config.enabledErrorTypes.OOMs)
                 && (lastBootInForeground && lastBootWasActive);
             [self deleteSentinelFile];
             return sameVersions && shouldReport;